requirements.txt
Django==4.2.20
pillow==11.1.0
pip==25.0.1
asgiref==3.8.1
sqlparse==0.5.3
djangorestframework==3.15.2
setuptools==58.1.0